Add animation methods to GraphicsComponent
